HTML and CSS: Bug fixes Ch01 - Ch08

+++ android-framework/src/extensions/delegate/fixed.ts
@@ -30,7 +30,7 @@ function reduceContainerWidth<T extends View>(node: T, value: string, offset: nu
         }
     }
     else if ($util.isLength(value)) {
-        const width = parseInt(value) - offset;
+        const width = parseFloat(value) - offset;
         if (width > 0) {
             return $util.formatPX(width);
         }
@@ -110,22 +110,13 @@ export default class Fixed<T extends View> extends squared.base.Extension<T> {
             const width = node.cssInitial('width', true);
             const minWidth = node.cssInitial('minWidth', true);
             if (node.documentBody && node.some(item => item.has('right'))) {
-                node.cssApply({
-                    width: 'auto',
-                    minWidth: 'auto'
-                }, true);
-                node.companion.cssApply({
-                    width,
-                    minWidth
-                }, true);
+                node.cssApply({ width: 'auto', minWidth: 'auto' }, true);
+                node.companion.cssApply({ width, minWidth }, true);
                 node.android('layout_width', 'match_parent');
             }
             else {
                 const offset = node.paddingLeft + node.paddingRight + (node.documentBody ? node.marginLeft + node.marginRight : 0);
-                node.companion.cssApply({
-                    width: reduceContainerWidth(node, width, offset),
-                    minWidth: reduceContainerWidth(node, minWidth, offset)
-                }, true);
+                node.companion.cssApply({ width: reduceContainerWidth(node, width, offset), minWidth: reduceContainerWidth(node, minWidth, offset) }, true);
             }
         }
     }
